+++ doc/whats_new.rst
@@ -118,6 +118,14 @@ Bug fixes
      when a numpy array is passed in for weights. :issue:`7983` by
      :user:`Vincent Pham <vincentpham1991>`.
 
+   - Fix a bug in :class:`sklearn.decomposition.LatentDirichletAllocation`
+     where the ``perplexity`` method was returning incorrect results because
+     the ``transform`` method retruns normalized document topic distributions
+     as of version 0.18. Also deprecated the ``doc_topic_distr`` argument of
+     the ``perplexity`` method because the user no longer has access to the
+     unnormalized document topic distribution needed for the perplexity
+     calculation. :issue:`7954` by :user:`Gary Foreman <garyForeman>`.
+
 API changes summary
 -------------------